ui/requirements.txt
streamlit>=1.33
httpx>=0.27
python-dotenv>=1.0
pydantic>=2.7
pandas>=2.2
plotly>=5.22